# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeMorph AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type JobType = typeof JobType[keyof typeof JobType];


export const JobType = {
  video_morph: 'video_morph',
  photo_animate: 'photo_animate',
} as const;

export type JobStatus = typeof JobStatus[keyof typeof JobStatus];


export const JobStatus = {
  queued: 'queued',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

/**
 * AI transformation style
 * @nullable
 */
export type JobStyle = typeof JobStyle[keyof typeof JobStyle] | null;


export const JobStyle = {
  anime: 'anime',
  cartoon: 'cartoon',
  ghibli: 'ghibli',
  cyberpunk: 'cyberpunk',
  watercolor: 'watercolor',
} as const;

export interface Job {
  id: number;
  userId: string;
  type: JobType;
  status: JobStatus;
  /**
     * User-supplied label for this job
     * @nullable
     */
  title?: string | null;
  /** Storage path of the uploaded input file */
  inputObjectPath: string;
  /**
     * Storage path of the processed output file
     * @nullable
     */
  outputObjectPath?: string | null;
  /**
     * Storage path of the thumbnail image
     * @nullable
     */
  thumbnailObjectPath?: string | null;
  /**
     * Processing progress 0–100
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  /** @nullable */
  errorMessage?: string | null;
  /**
     * AI transformation style
     * @nullable
     */
  style?: JobStyle;
  /**
     * Duration of the input video in seconds
     * @nullable
     */
  durationSeconds?: number | null;
  createdAt: string;
  updatedAt: string;
}

export interface JobList {
  jobs: Job[];
  total: number;
}

export type JobInputType = typeof JobInputType[keyof typeof JobInputType];


export const JobInputType = {
  video_morph: 'video_morph',
  photo_animate: 'photo_animate',
} as const;

export type JobInputStyle = typeof JobInputStyle[keyof typeof JobInputStyle];


export const JobInputStyle = {
  anime: 'anime',
  cartoon: 'cartoon',
  ghibli: 'ghibli',
  cyberpunk: 'cyberpunk',
  watercolor: 'watercolor',
} as const;

export interface JobInput {
  type: JobInputType;
  /** Optional user-supplied label */
  title?: string;
  /** Storage path returned by the upload URL endpoint */
  inputObjectPath: string;
  style?: JobInputStyle;
}

export interface JobStats {
  total: number;
  queued: number;
  processing: number;
  completed: number;
  failed: number;
  videoMorphCount: number;
  photoAnimateCount: number;
}

export interface UploadUrlRequest {
  /**
     * Original file name.
     * @minLength 1
     */
  name: string;
  /**
     * File size in bytes.
     * @minimum 1
     */
  size: number;
  /**
     * MIME type of the file.
     * @minLength 1
     */
  contentType: string;
}

export interface UploadUrlResponse {
  /** Presigned GCS URL for PUT upload. */
  uploadURL: string;
  /** Normalized object path. Store this in your database. */
  objectPath: string;
  metadata?: UploadUrlRequest;
}

export interface ErrorEnvelope {
  error: string;
}

export type ListJobsParams = {
/**
 * Filter by job status
 */
status?: ListJobsStatus;
/**
 * Filter by job type
 */
type?: ListJobsType;
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
/**
 * @minimum 0
 */
offset?: number;
};

export type ListJobsStatus = typeof ListJobsStatus[keyof typeof ListJobsStatus];


export const ListJobsStatus = {
  queued: 'queued',
  processing: 'processing',
  completed: 'completed',
  failed: 'failed',
} as const;

export type ListJobsType = typeof ListJobsType[keyof typeof ListJobsType];


export const ListJobsType = {
  video_morph: 'video_morph',
  photo_animate: 'photo_animate',
} as const;