# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AnimeMorph AI API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: Jobs
    description: Render job management
  - name: Storage
    description: Object storage upload and serving endpoints

paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /jobs:
    get:
      operationId: listJobs
      tags: [Jobs]
      summary: List all render jobs for the current user
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [queued, processing, completed, failed]
          description: Filter by job status
        - name: type
          in: query
          required: false
          schema:
            type: string
            enum: [video_morph, photo_animate]
          description: Filter by job type
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: List of jobs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    post:
      operationId: createJob
      tags: [Jobs]
      summary: Submit a new render job
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobInput"
      responses:
        "201":
          description: Job created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /jobs/stats:
    get:
      operationId: getJobStats
      tags: [Jobs]
      summary: Get dashboard statistics for the current user
      responses:
        "200":
          description: Job statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/JobStats"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /jobs/recent:
    get:
      operationId: getRecentJobs
      tags: [Jobs]
      summary: Get the 5 most recent jobs across all statuses
      responses:
        "200":
          description: Recent jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Job"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /jobs/{id}:
    get:
      operationId: getJob
      tags: [Jobs]
      summary: Get a specific render job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    delete:
      operationId: deleteJob
      tags: [Jobs]
      summary: Delete a render job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /jobs/{id}/retry:
    post:
      operationId: retryJob
      tags: [Jobs]
      summary: Retry a failed render job
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Job queued for retry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Job is not in a failed state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
        "404":
          description: Job not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"

  /storage/uploads/request-url:
    post:
      tags: [Storage]
      operationId: requestUploadUrl
      summary: Request a presigned URL for file upload
      description: |
        Returns a presigned GCS URL for direct upload. The client sends JSON
        metadata here, then uploads the file directly to the returned URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UploadUrlRequest'
      responses:
        '200':
          description: Presigned upload URL generated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadUrlResponse'
        '400':
          description: Missing or invalid required fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to generate upload URL.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /storage/public-objects/{filePath}:
    get:
      tags: [Storage]
      operationId: getPublicObject
      summary: Serve a public asset from PUBLIC_OBJECT_SEARCH_PATHS
      parameters:
        - name: filePath
          in: path
          required: true
          schema:
            type: string
          description: Relative file path within the public search paths.
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: File not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to serve public object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

  /storage/objects/{objectPath}:
    get:
      tags: [Storage]
      operationId: getStorageObject
      summary: Serve an object entity from PRIVATE_OBJECT_DIR
      parameters:
        - name: objectPath
          in: path
          required: true
          schema:
            type: string
          description: Object path within the private object dir.
      responses:
        '200':
          description: Object content streamed with correct Content-Type.
          content:
            '*/*':
              schema:
                type: string
                format: binary
        '404':
          description: Object not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Failed to serve object.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Job:
      type: object
      required: [id, userId, type, status, inputObjectPath, progress, createdAt, updatedAt]
      properties:
        id:
          type: integer
        userId:
          type: string
        type:
          type: string
          enum: [video_morph, photo_animate]
        status:
          type: string
          enum: [queued, processing, completed, failed]
        title:
          type: ["string", "null"]
          description: User-supplied label for this job
        inputObjectPath:
          type: string
          description: Storage path of the uploaded input file
        outputObjectPath:
          type: ["string", "null"]
          description: Storage path of the processed output file
        thumbnailObjectPath:
          type: ["string", "null"]
          description: Storage path of the thumbnail image
        progress:
          type: integer
          minimum: 0
          maximum: 100
          description: Processing progress 0–100
        errorMessage:
          type: ["string", "null"]
        style:
          type: ["string", "null"]
          enum: [anime, cartoon, ghibli, cyberpunk, watercolor, null]
          description: AI transformation style
        durationSeconds:
          type: ["number", "null"]
          description: Duration of the input video in seconds
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    JobList:
      type: object
      required: [jobs, total]
      properties:
        jobs:
          type: array
          items:
            $ref: "#/components/schemas/Job"
        total:
          type: integer

    JobInput:
      type: object
      required: [type, inputObjectPath]
      properties:
        type:
          type: string
          enum: [video_morph, photo_animate]
        title:
          type: string
          description: Optional user-supplied label
        inputObjectPath:
          type: string
          description: Storage path returned by the upload URL endpoint
        style:
          type: string
          enum: [anime, cartoon, ghibli, cyberpunk, watercolor]
          default: anime

    JobStats:
      type: object
      required: [total, queued, processing, completed, failed, videoMorphCount, photoAnimateCount]
      properties:
        total:
          type: integer
        queued:
          type: integer
        processing:
          type: integer
        completed:
          type: integer
        failed:
          type: integer
        videoMorphCount:
          type: integer
        photoAnimateCount:
          type: integer

    UploadUrlRequest:
      type: object
      required: [name, size, contentType]
      properties:
        name:
          type: string
          minLength: 1
          description: Original file name.
        size:
          type: integer
          minimum: 1
          description: File size in bytes.
        contentType:
          type: string
          minLength: 1
          description: MIME type of the file.

    UploadUrlResponse:
      type: object
      required: [uploadURL, objectPath]
      properties:
        uploadURL:
          type: string
          format: uri
          description: Presigned GCS URL for PUT upload.
        objectPath:
          type: string
          description: Normalized object path. Store this in your database.
        metadata:
          $ref: '#/components/schemas/UploadUrlRequest'

    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string