# lib/api-zod/src/generated/types/job.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeMorph AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobStatus } from './jobStatus';
import type { JobStyle } from './jobStyle';
import type { JobType } from './jobType';

export interface Job {
  id: number;
  userId: string;
  type: JobType;
  status: JobStatus;
  /**
     * User-supplied label for this job
     * @nullable
     */
  title?: string | null;
  /** Storage path of the uploaded input file */
  inputObjectPath: string;
  /**
     * Storage path of the processed output file
     * @nullable
     */
  outputObjectPath?: string | null;
  /**
     * Storage path of the thumbnail image
     * @nullable
     */
  thumbnailObjectPath?: string | null;
  /**
     * Processing progress 0–100
     * @minimum 0
     * @maximum 100
     */
  progress: number;
  /** @nullable */
  errorMessage?: string | null;
  /**
     * AI transformation style
     * @nullable
     */
  style?: JobStyle;
  /**
     * Duration of the input video in seconds
     * @nullable
     */
  durationSeconds?: number | null;
  createdAt: Date;
  updatedAt: Date;
}