# lib/api-zod/src/generated/types/jobInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeMorph AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { JobInputStyle } from './jobInputStyle';
import type { JobInputType } from './jobInputType';

export interface JobInput {
  type: JobInputType;
  /** Optional user-supplied label */
  title?: string;
  /** Storage path returned by the upload URL endpoint */
  inputObjectPath: string;
  style?: JobInputStyle;
}