# lib/api-zod/src/generated/types/jobList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeMorph AI API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Job } from './job';

export interface JobList {
  jobs: Job[];
  total: number;
}