# lib/api-zod/src/generated/types/jobStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * AnimeMorph AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface JobStats {
  total: number;
  queued: number;
  processing: number;
  completed: number;
  failed: number;
  videoMorphCount: number;
  photoAnimateCount: number;
}